+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix bug where new messages were not added to the list on refresh if skip_deleted=true (#8234)
 - Fix colors on "Show source" page in dark mode (#8246)
 - Fix handling of dark_mode_support:false setting in skins meta.json - also when devel_mode=false (#8249)
+- Fix database initialization if db_prefix is a schema prefix (#8221)
 
 ## Release 1.5.0